Update by SupportsArrayIndexing[T_co]
